backend/requirements.txt
# 车牌识别系统依赖
# Python 3.9+

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.15.0

# 计算机视觉
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# 目标检测 (YOLOv8)
ultralytics>=8.0.0

# OCR相关
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# 中国车牌识别专用库
hyperlpr3>=0.1.0

# Web框架
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# 数据验证
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 配置管理
pyyaml>=6.0

# 日志
loguru>=0.7.0

# 视频处理
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# 工具库
python-dotenv>=1.0.0
aiofiles>=23.0.0
httpx>=0.24.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
